+++ fault-localization-toolset/src/main/java/se/de/hu_berlin/informatik/faultlocalizer/sbfl/localizers/Kulczynski1.java
@@ -33,7 +33,7 @@ public class Kulczynski1<T> extends AbstractSpectrumBasedFaultLocalizer<T> {
     	if (node.getEF(strategy) == 0) {
     		return 0;
     	}
-        return (double)node.getEF(strategy) / (double)(node.getNF(strategy) + node.getEP(strategy));
+        return node.getEF(strategy) / (node.getNF(strategy) + node.getEP(strategy));
     }
 
     @Override
